Rename all manager modules add `_manager` suffix
