+++ rust/src/manager.rs
@@ -1,6 +1,7 @@
-pub mod auth;
+pub mod auth_manager;
+pub mod coin_control_manager;
 pub mod deferred_sender;
-pub mod import_wallet;
-pub mod pending_wallet;
+pub mod import_wallet_manager;
+pub mod pending_wallet_manager;
 pub mod send_flow_manager;
-pub mod wallet;
+pub mod wallet_manager;
