+++ rust/src/manager/auth_manager.rs
