+++ rust/src/manager/pending_wallet_manager.rs
