+++ rust/src/manager/wallet_manager.rs
